egg.TreeDictionary autosettings now automatic!
What a hassle. Passes all linux tests.

+++ tests/test__egg.py
@@ -29,7 +29,7 @@ class Test_egg(_ut.TestCase):
         Just opens a full-featured example and lets the user play with it
         """
         import spinmob.egg.example_sweeper as sweeper
-        _sh.rmtree('gui_settings')
+        if _os.path.exists('gui_settings'): _sh.rmtree('gui_settings')
         
         sweeper.d_sweep.load_file(_os.path.join(self.data_path, 'difficult.binary'))
         
@@ -42,17 +42,23 @@ class Test_egg(_ut.TestCase):
         """
         import spinmob.egg as _e
         
+        # clear out the gui_settings
+        if _os.path.exists('gui_settings'): _sh.rmtree('gui_settings')
+        
         test_list_values = [42, 'test', 'pants', 37.2, dict(stuff='otherstuff')]
         test_list_values_stringified = [str(42), 'test', 'pants', str(37.2), str(dict(stuff='otherstuff'))]
         
         # Create the TreeDictionary
-        self.d = _e.gui.TreeDictionary()
+        self.d = _e.gui.TreeDictionary('pants.txt')
+        
+        # Make sure the autosettings path is right.
+        self.assertEqual(self.d._autosettings_path, 'pants.txt')
         
         # Create some different value types
         self.d.add_button('button')
-        self.d.add_parameter('booly',   value=False,  type='bool')
-        self.d.add_parameter('inty',    value=42,     type='int')
-        self.d.add_parameter('floaty',  value=42.0,   type='float')
+        self.d.add_parameter('booly',   value=False)
+        self.d.add_parameter('inty',    value=42)
+        self.d.add_parameter('floaty',  value=42.0)
         self.d.add_parameter('stringy', value=574,    type='str')
         self.d.add_parameter('listy',   value='test', type='list', values=test_list_values)
     
@@ -74,6 +80,9 @@ class Test_egg(_ut.TestCase):
         self.assertEqual(type(self.d['floaty']), float)
         self.d['floaty'] = '45.5555'
         self.assertEqual(self.d['floaty'], 45.5555)
+
+        # See if it's autosaving
+        self.assertTrue(_os.path.exists('gui_settings/pants.txt'))
         
         # String stuff
         self.assertEqual(type(self.d['stringy']), str)
@@ -89,10 +98,20 @@ class Test_egg(_ut.TestCase):
         # Save, load, and make sure the values are still the same (with types)
         self.d.save()
         self.d.load()
+        self.assertTrue(_os.path.exists('gui_settings/pants.txt'))
         self.assertEqual(self.d['inty'], 32)
         self.assertEqual(self.d['floaty'], 45.5555)
         self.assertEqual(type(self.d['stringy']),  str)
         self.assertEqual(self.d['listy'], str(37.2))
+        
+        # Now change a parameter, which should autosave
+        self.d['floaty'] = 32
+        
+        # Create a new one with the same autosettings path
+        self.e = _e.gui.TreeDictionary(autosettings_path='pants.txt')
+        self.e.add_parameter('floaty', 42.0)
+        self.assertEqual(self.e['floaty'],32)
+        
         
 
 if __name__ == "__main__":
